+++ CHANGELOG.md
@@ -11,17 +11,20 @@ Most recent change on the bottom.
 ### Added
 - `NequIPCalculator` now handles per-atom energies
 - Added `initial_model_state_strict` YAML option
+- `load_model_state` builder
 - fusion strategy support
 - multiprocessing for ASE dataset loading/processing
 
 ### Changed
 - Disallow PyTorch 1.9, which has some JIT bugs.
+- `nequip-deploy build` now requires `--train-dir` option when specifying the training session
 
 ### Fixed
 - Better error in `Dataset.statistics` when field is missing
 - `NequIPCalculator` now outputs energy as scalar rather than `(1, 1)` array
 - `dataset: ase` now treats automatically adds `key_mapping` keys to `include_keys`, which is consistant with the npz dataset
 - fixed reloading models with `per_species_rescale_scales/shifts` set to `null`/`None`
+- graceful exit for `-n 0` in `nequip-benchmark`
 
 ## [0.5.3] - 2022-02-23
 ### Added